refactor(builder): move the breakpoint editor where its rules can be derived
packages/ui is the block-agnostic layer, so it cannot depend on the block
engine. lib/breakpoints.ts therefore mirrored the engine's breakpoint types
structurally and kept its own copy of MAX_BREAKPOINTS_PER_AXIS, restating the
drop rules the style compiler applies. Two implementations of one rule agree
the day they are written and drift silently after; the ui layering guard named
this as the reason that package was not yet block-agnostic.

Both files now live in packages/builder, which already depends on the engine
and imports the cap and the types rather than mirroring them.

The per-rule drop decisions still mirror compile-page.ts, which makes them
inline while emitting, so there is no predicate to call. Exporting one from the
engine is the remaining half and is an engine change; the module says so.

The @nextlyhq/ui/breakpoints subpath, BreakpointDialog and the breakpoint types
leave the ui export map. Nothing in the repository imported them and every
affected export was experimental.
