+++ packages/builder/package.json
@@ -46,6 +46,7 @@
   "peerDependencies": {
     "@nextlyhq/plugin-sdk": "workspace:*",
     "@nextlyhq/ui": "workspace:^",
+    "lucide-react": ">=0.400.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0"
   },
@@ -55,10 +56,13 @@
     "@nextlyhq/plugin-sdk": "workspace:*",
     "@nextlyhq/tsconfig": "workspace:*",
     "@nextlyhq/ui": "workspace:*",
+    "@testing-library/react": "^16.3.0",
     "@types/node": "^20.19.17",
     "@types/react": "19.2.0",
     "@types/react-dom": "19.2.0",
     "eslint": "^9.39.1",
+    "jsdom": "^27.1.0",
+    "lucide-react": "^0.544.0",
     "react": "19.2.0",
     "react-dom": "19.2.0",
     "rimraf": "^6.1.3",
